Remove leaflet from hazdev-location-view
