--- projects/hazdev-ng-location-view/src/lib/location-map/location-map.component.css
@@ -1,16 +1,0 @@
-div.map {
-  height: 450px;
-}
-
-.leaflet-location-control {
-  cursor: pointer;
-}
-
-.leaflet-location-control > a.material-icons {
-  font-size: 16px;
-  font-weight: bold;
-}
-
-.mat-dialog-container {
-  background: #fafafa;
-}
