--- projects/hazdev-ng-location-view/src/lib/location-map/location-map.component.html
@@ -1,4 +1,0 @@
-<div #mapWrapper
-    class="map"
-    (mousemove)="$event.stopPropagation">
-</div>
