+++ projects/hazdev-ng-location-view/src/public_api.ts
@@ -12,4 +12,3 @@ export * from './lib/coordinate-input/coordinate-input.component';
 export * from './lib/geocode-input/geocode-input.component';
 export * from './lib/geolocate-input/geolocate-input.component';
 export * from './lib/location-dialog/location-dialog.component';
-export * from './lib/location-map/location-map.component';
