+++ src/app/app.component.scss
@@ -1,0 +1,12 @@
+div.map {
+  height: 450px;
+}
+
+.leaflet-location-control {
+  cursor: pointer;
+}
+
+.leaflet-location-control > a.material-icons {
+  font-size: 16px;
+  font-weight: bold;
+}
